Fixed Implement /api/update-hunt-begin issue in MatchQuestionnaireActivity #137

+++ app/src/main/java/com/recep/hunt/home/adapter/NearByRestaurantsAdapterByApi.kt
@@ -214,7 +214,7 @@ class NearByRestaurantsAdapterByApi(
                                     result[i].basicInfo.job_title,
                                     result[i].basicInfo.about,
                                     result[i].totalMatching,
-                                    result[i].totalMeeting,
+                                    result[i].totalMeeting.toInt(),
                                     result[i].is_online,
                                     result[i].for_date,
                                     result[i].for_bussiness,
